requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.1.3
django-restframework==0.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et_xmlfile==2.0.0
openpyxl==3.1.5
pillow==11.0.0
PyJWT==2.9.0
reportlab==4.2.5
sqlparse==0.5.1
